+++ idea-plugin/src/main/kotlin/com/flowable/atlas/settings/GenerationLiquibaseConfigurable.kt
@@ -21,6 +21,7 @@ class GenerationLiquibaseConfigurable(project: Project) : AtlasProjectConfigurab
     override fun createPanel(): DialogPanel {
         val settings = FlowableAtlasProjectSettings.getInstance(project)
         return panel {
+            scopeLine(project)
             row("Output folder:") {
                 textFieldWithBrowseButton(
                     FileChooserDescriptorFactory.createSingleFolderDescriptor()